Improve checkbbox
